requirements.txt
pyyaml>=6.0
requests>=2.31.0
nltk>=3.8.1

